Finished the light mode feature, made it so that the preference is saved when changed and loaded on startup.
