+++ src/PixiEditor/Data/Localization/Languages/en.json
@@ -1,5 +1,6 @@
 {
   "RECENT_FILES": "Recent Files",
+  "CHANGE_THEME": "Theme",
   "OPEN_FILE": "Open file",
   "NEW_FILE": "New",
   "RECENT_EMPTY_TEXT": "So much empty space",
