+++ src/PixiEditor/ViewModels/UserPreferences/Settings/SceneSettings.cs
@@ -69,22 +69,9 @@ internal class SceneSettings : SettingsGroup
 
     private string requestedTheme = GetPreference(PreferencesConstants.RequestedTheme, PreferencesConstants.RequestedThemeDefault);
 
-    private ThemeVariant convertThemeName(string requestedTheme)
-    {
-        ThemeVariant variant = requestedTheme switch
-        {
-            "Dark" => ThemeVariant.Dark,
-            "Light" => ThemeVariant.Light,
-            "Default" => ThemeVariant.Default,
-            _ => ThemeVariant.Default
-        };
-        return variant;
-    }
-
     public string RequestedTheme
     {
         get => requestedTheme;
-        //set => App.Current.RequestedThemeVariant = convertThemeName(requestedTheme);
         set => RaiseAndUpdatePreference(PixiEditorSettings.Scene.RequestedTheme, value);
     }
 
@@ -95,7 +82,8 @@ internal class SceneSettings : SettingsGroup
             PrimaryBackgroundColorHex = PreferencesConstants.PrimaryBackgroundColorDefault;
             SecondaryBackgroundColorHex = PreferencesConstants.SecondaryBackgroundColorDefault;
         });
-        
+
+        SubscribeSettingsValueChanged(PixiEditorSettings.Scene.RequestedTheme, nameof(RequestedTheme));
         SubscribeValueChanged(PixiEditorSettings.Tools.SelectionTintingEnabled, nameof(SelectionTintingEnabled));
     }
 }
